fix(engine): combine migrations
Combines the ProcessDefinitionVersionMigration with the
MultiTenancyMigration. This way we can profit from the needsToRun check
of the MultiTenancyMigration. It's also more efficient performance-wise,
 as we only need to go through the CF one time, instead of two.
