+++ engine/src/main/java/io/camunda/zeebe/engine/state/migration/DbMigratorImpl.java
@@ -10,7 +10,6 @@ package io.camunda.zeebe.engine.state.migration;
 import io.camunda.zeebe.engine.state.migration.to_8_2.DecisionMigration;
 import io.camunda.zeebe.engine.state.migration.to_8_2.DecisionRequirementsMigration;
 import io.camunda.zeebe.engine.state.migration.to_8_3.MultiTenancyMigration;
-import io.camunda.zeebe.engine.state.migration.to_8_3.ProcessDefinitionVersionMigration;
 import io.camunda.zeebe.engine.state.migration.to_8_3.ProcessInstanceByProcessDefinitionMigration;
 import io.camunda.zeebe.engine.state.mutable.MutableProcessingState;
 import java.util.ArrayList;
@@ -34,7 +33,6 @@ public class DbMigratorImpl implements DbMigrator {
           new DecisionMigration(),
           new DecisionRequirementsMigration(),
           new ProcessInstanceByProcessDefinitionMigration(),
-          new ProcessDefinitionVersionMigration(),
           new MultiTenancyMigration());
   // Be mindful of https://github.com/camunda/zeebe/issues/7248. In particular, that issue
   // should be solved first, before adding any migration that can take a long time
