+++ engine/src/test/java/io/camunda/zeebe/engine/state/migration/to_8_3/legacy/LegacyProcessState.java
@@ -422,5 +422,12 @@ public final class LegacyProcessState {
     public void clear() {
       versionCache.clear();
     }
+
+    public void insertProcessVersion(final String processId, final int version) {
+      processIdKey.wrapString(processId);
+      final var value = new NextValue();
+      value.set(version);
+      nextValueColumnFamily.insert(processIdKey, value);
+    }
   }
 }
